test(handler): isolate handler error-response tests in their own integration binary
The four handler script-failure contract tests spin up real listeners running
inject/decorate scripts. Co-located in the rift-core lib test binary they shared
the process-global MB JS script pool with the timing-sensitive bounded-matcher
unit tests, whose runaway-script cases park pool workers until the loop cap throws;
under CI load that starved the bounded inject matcher's 60s deadline and failed CI.

Move them to a dedicated tests/handler_error_responses.rs integration binary (a
separate process with its own script pool), and run it via a dedicated CI step
since the main '--lib' job excludes tests/*.rs. The pure matcher_error_response
branch unit tests stay in the lib.
